refactoring and modularizing files
